9th question company salary bonus....
The question says that if the employee is male then he will get 5% bonus on his salary. If the employee's salary is less than 10000 then the employee will get 2% additional bonus on his salary. 
It means that if the employee is male then he will get 5% bonus on his salary and if the salary is less than 10000 then he will get 2% additional bonus on his salary.
If the employee is not male then she will get only 2% bonus on his salary.
But According to your program if the employee is male then he will get 5% bonus on his salary.
else she will get 10% bonus on his salary.
So, that's the mistake in your program hope you understand the logic. 👍 :)